REFACTOR: magical permissions is inside magical camera
